backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
httpx
python-dotenv
numpy
pandas
scikit-learn
matplotlib
seaborn
plotly
apscheduler
instaloader